Completely refactor project from python to go
**Why am I doing this?**
Well, I'm doing this because i will eternally despise python. I
initially used it because i thought this would be a small project, and
python is great for small projects!

<helloworld.py>
    print("Hello World!")

<helloworld.go>
    package main

    import (
        "fmt"
    )

    func main() {
        fmt.Println("Hello World!")
    }

As you can see, much less code. But in bigger projects; you just got to
accept it:

Go is better.

It just is. You require much less code, and it's faster! So yeah, that's
why I'm doing this.

**Changes**
[README.RST] <- Modify to include the Go requirements.
[FMAKE->FMAKE.GO] <- Refactor to be comprised of Go code.
[UTILS/UTILS.GO] <- Main module that is called by the main executable.
[MAKEFILE] <- Add to compile Go code.